pyproject.toml
[tool.poetry]
name = "logo-placement-background-remover"
version = "0.1.0"
description = "variphi codeing assignment \"In this assigment I want to do the Logo placement and also BG remover\""
authors = ["Darshan A M"]
license = "None"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
sqlalchemy = "^2.0.32"
pillow = "^10.4.0"
opencv-python = "^4.10.0.84"
pydantic = "^2.8.2"
torch = "^2.4.0"
transformers = "^4.44.2"
diffusers = "^0.30.2"
scikit-image = "^0.24.0"
torchsde = "^0.2.6"
torchdiffeq = "^0.2.4"
tomesd = "^0.1.3"
safetensors = "^0.4.4"
open-clip-torch = "^2.26.1"
torchvision = "^0.19.0"
pytorch-lightning = "^2.4.0"
segment-anything = {git = "https://github.com/facebookresearch/segment-anything.git"}
python-dotenv = "^1.0.1"
pydantic-settings = "^2.4.0"
psutil = "^6.0.0"
gunicorn = "^23.0.0"
mysql-connector-python = "^9.0.0"
scikit-learn = "^1.5.1"
numpy = "^2.1.1"
matplotlib = "^3.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
